Add benchmarks, test vectors, multidevice features, and CI workflow
New features:
- Account recovery module for trelis-hybrid
- Device approval certificates and key wrapping for multidevice
- Memory locking (mlock) for secret protection in primitives

Test infrastructure:
- Comprehensive test vectors for x3dh-pq, ratchet, safety-number, hybrid-sig, cocoa, multidevice
- Error path testing across all crates
- Multidevice integration tests
- GitHub Actions CI workflow

Performance:
- Benchmarks for CoCoA operations (tree ops, encryption, init)
- Benchmarks for hybrid primitives (KEM, signatures)
- Benchmarks for ratchet messaging

Build improvements:
- Switch flake.nix toolchain from nightly to stable
- Remove nightly-only options from rustfmt.toml for stable compatibility
- Apply consistent formatting across all crates

+++ crates/trelis-cocoa/src/tree/index.rs
@@ -132,7 +132,7 @@ impl NodeIndex {
         }
 
         // Find depth: depth d starts at index 2^d - 1
-        let depth = (64 - (linear + 1).leading_zeros() - 1) as u32;
+        let depth = 64 - (linear + 1).leading_zeros() - 1;
         let nodes_before = (1u64 << depth) - 1;
         let position = (linear - nodes_before) as u32;
